--- demo/src/main/java/com/example/demo/Auth/LoginRequest.java
@@ -1,6 +1,0 @@
-package com.example.demo.Auth;
-
-public class LoginRequest {
-    public String email;
-    public String password;
-}
