+++ demo/src/main/java/com/example/demo/Auth/Service.java
@@ -1,47 +1,45 @@
 package com.example.demo.Auth;
-
 import java.util.ArrayList;
+    @org.springframework.stereotype.Service
+    public class Service {
+        ArrayList<User> DB = new ArrayList<>();
+        public Boolean login(Loginrequest loginRequest) {
+            String email = loginRequest.email;
+            String password = loginRequest.password;
 
-@org.springframework.stereotype.Service
-public class Service {
-    ArrayList<User> DB = new ArrayList<>();
-    public Boolean login(LoginRequest loginRequest) {
-        String email = loginRequest.email;
-        String password = loginRequest.password;
+            User user  = findUserByEmail(email);
+            if(user == null){
+                return false;
+            }
 
-        User user  = findUserByEmail(email);
-        if(user == null){
+            if(user.getPassword().equals(password)){
+                return true;
+            }
             return false;
         }
 
-        if(user.getPassword().equals(password)){
+        private User findUserByEmail(String email) {
+            for(User user : DB){
+                if(user.getEmail().equals(email)){
+                    return user;
+                }
+            }
+            return null;
+        }
+
+        public Boolean signup(Signuprequest signupRequest) {
+            String email = signupRequest.email;
+            User user = findUserByEmail(email);
+            if(user != null){
+                return false;
+            }
+
+            User newUser = new User(signupRequest.name, signupRequest.email, signupRequest.password);
+            saveUser(newUser);
             return true;
         }
-        return false;
-    }
 
-    private User findUserByEmail(String email) {
-        for(User user : DB){
-            if(user.getEmail().equals(email)){
-                return user;
-            }
+        private void saveUser(User newUser) {
+            DB.add(newUser);
         }
-        return null;
     }
-
-    public Boolean signup(SignupRequest signupRequest) {
-        String email = signupRequest.email;
-        User user = findUserByEmail(email);
-        if(user != null){
-            return false;
-        }
-
-        User newUser = new User(signupRequest.name, signupRequest.email, signupRequest.password);
-        saveUser(newUser);
-        return true;
-    }
-
-    private void saveUser(User newUser) {
-        DB.add(newUser);
-    }
-}
