+++ demo/src/main/java/com/example/demo/Auth/Signuprequest.java
@@ -1,6 +1,6 @@
 package com.example.demo.Auth;
 
-public class SignupRequest {
+public class Signuprequest {
     public String name;
     public String email;
     public String password;
