+++ demo/src/main/java/com/example/demo/Auth/Userrepository.java
@@ -1,0 +1,6 @@
+package com.example.demo.Auth;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface Userrepository extends MongoRepository {
+}
