--- demo/src/main/java/com/example/demo/Signuprequest.java
@@ -1,8 +1,0 @@
-package com.example.demo;
-
-public class Signuprequest {
-    public int enrollno;
-    public String name;
-
-    public String password;
-}
