+++ demo/src/main/java/com/example/demo/Student.java
@@ -2,23 +2,23 @@ package com.example.demo;
 
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.annotation.Id;
-@Document(collection = "Student")
 public class Student {
     @Id
     private int enrollno;
     private String name;
     private String branch;
     private String password;
-    public Student(String name, int enrollno, String password){
 
-    }
-
-    public Student(int enrollno, String name, String branch,String Password) {
+    public Student(int enrollno, String name, String branch,String password) {
 
         this.enrollno = enrollno;
         this.name = name;
         this.branch = branch;
         this.password=password;
+    }
+
+    public Student(int enrollno, String name,String password) {
+
     }
 
     public void setEnrollno(int enrollno) {
@@ -39,11 +39,12 @@ public class Student {
     public String getBranch() {
         return branch;
     }
-    public String getPassword(){
-        return password;
-    }
+
     public void setPassword(String password) {
         this.password = password;
+    }
+    public String getPassword() {
+        return password;
     }
 
 }
