+++ demo/src/main/java/com/example/demo/StudentController.java
@@ -33,27 +33,6 @@ public class StudentController {
 
         return "Deleted Successfully";
     }
-    @Autowired
-    Service service;
-    @PostMapping(path = "login")
-    public String login(@RequestBody Loginrequest loginRequest){
-        Boolean loggedIn = service.login(loginRequest);
-        if(loggedIn)return "Logged In with email "+ loginRequest.enrollno;
-        else{
-            return "Please check your email and password";
-        }
-    }
-    @PostMapping(path = "signup")
-    public String signup(@RequestBody Signuprequest signupRequest){
-        Boolean signedUp = service.signup(signupRequest);
-        if(signedUp)return "Created a new user with email "+ signupRequest.enrollno;
-        else{
-            return "User with this email already exists";
-        }
-    }
-
-
-
 
 }
